changed api for jokes and implimented a generate token
