+++ config/middlewares.js
@@ -5,7 +5,21 @@ const jwtKey = require('../_secrets/keys').jwtKey;
 // quickly see what this file exports
 module.exports = {
   authenticate,
+  generateToken,
 };
+
+function generateToken(user) {
+  const payload = {
+      subject: user.id,
+      username: user.username,
+  };
+
+  const options = {
+      expiresIn: '1h',
+  };
+
+  return jwt.sign(payload, jwtKey, options);
+}
 
 // implementation details
 function authenticate(req, res, next) {
